Police Shooting in Seattle, WA data visualization
This is a report of a police shooting in the city of Seattle, Washington. The data gives details about the number of years in the police service, the background information about the police and the subject, how many rounds were fired, as well as if the shooting is fatal or not. While exploring this data, I did some data wrangling and data visualization to explore how different variables compare to each data.